REFACTOR: change model module in chat persistence, core layer
