+++ chat/core/build.gradle.kts
@@ -11,6 +11,7 @@ repositories {
 
 dependencies {
     implementation(project(":chat:persistence"))
+    implementation(project(":model"))
 
     //WEBFLUX
     implementation("org.springframework.boot:spring-boot-starter-webflux")
