+++ chat/persistence/build.gradle.kts
@@ -10,6 +10,9 @@ repositories {
 }
 
 dependencies {
+    //MODULE
+    implementation(project(":model"))
+
     //WEBFLUX
     implementation("org.springframework.boot:spring-boot-starter-webflux")
