+++ chat/persistence/src/main/kotlin/vp/togedo/repository/ChatRepository.kt
@@ -3,7 +3,7 @@ package vp.togedo.repository
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
 import reactor.core.publisher.Flux
-import vp.togedo.document.ChatDocument
+import vp.togedo.model.documents.chat.ChatDocument
 
 interface ChatRepository: ReactiveMongoRepository<ChatDocument, ObjectId> {
